Multitasking: some changes and fixes in tasks and programs - add impl of Task_GetTaskAddress for program.c - add test syscalls for Task_Create and Task_Launch - change in Task_Free: no more Task_Kill call, call it before Task_Free yourself - split code from context.S into context.S and manager.S

+++ programs/programs.h
@@ -21,6 +21,6 @@ typedef struct
 extern Program_Item programs[PROGRAMS_MAX_COUNT];
 
 int32_t Program_Create(void*(*func)(void*), void* arg);
-int32_t Program_Terminate(int32_t id);
+int32_t Program_Terminate(uint16_t id);
 
 #endif //TIOO_PROGRAMS_H_
